// src/ifcu/ifcu.enums.ts
export enum IFCU_CMD {
  IFCU_ON = 'IFCU_ON',
  IFCU_OFF = 'IFCU_OFF',
  IFCU_MODE_AUTO_COOL = 'IFCU_MODE_AUTO_COOL',
  IFCU_MODE_MANUAL_COOL = 'IFCU_MODE_MANUAL_COOL',
  IFCU_MODE_FAN_ONLY = 'IFCU_MODE_FAN_ONLY',
  IFCU_MODE_AUTO_HEAT = 'IFCU_MODE_AUTO_HEAT',
  IFCU_MODE_MANUAL_HEAT = 'IFCU_MODE_MANUAL_HEAT',
  IFCU_FAN_SPEED_LOW = 'IFCU_FAN_SPEED_LOW',
  IFCU_FAN_SPEED_MEDIUM = 'IFCU_FAN_SPEED_MEDIUM',
  IFCU_FAN_SPEED_HIGH = 'IFCU_FAN_SPEED_HIGH',
  IFCU_FAN_SPEED_TEN_SPEED = 'IFCU_FAN_SPEED_TEN_SPEED',
  IFCU_ACTION_INCREASE_TEMP = 'IFCU_ACTION_INCREASE_TEMP',
  IFCU_ACTION_DECREASE_TEMP = 'IFCU_ACTION_DECREASE_TEMP',
}

export enum SYSTEM_MODE {
  MANUAL = 'MANUAL',
  AUTO = 'AUTO',
}

export enum WARNING_LEVEL {
  AMBER = 'AMBER',
  RED = 'RED',
  BLACK = 'BLACK',
}

export enum ACTION_CODE {
  ISSUE = 'ISSUE',
  CANCEL = 'CANCEL',
}

export enum WARNING_CODE {
  HSWW_AMBER_OUTDOOR = 'HSWW_AMBER_OUTDOOR',
  HSWW_RED_OUTDOOR = 'HSWW_RED_OUTDOOR',
  HSWW_BLACK_OUTDOOR = 'HSWW_BLACK_OUTDOOR',
  HSWW_CANCEL_OUTDOOR = 'HSWW_CANCEL_OUTDOOR',

  HSWW_AMBER_INDOOR = 'HSWW_AMBER_INDOOR',
  HSWW_RED_INDOOR = 'HSWW_RED_INDOOR',
  HSWW_BLACK_INDOOR = 'HSWW_BLACK_INDOOR',
  HSWW_CANCEL_INDOOR = 'HSWW_CANCEL_INDOOR',
}

export const DO_MAPPING = {
  [WARNING_CODE.HSWW_AMBER_OUTDOOR]: 128,
  [WARNING_CODE.HSWW_RED_OUTDOOR]: 64,
  [WARNING_CODE.HSWW_BLACK_OUTDOOR]: 32,
  [WARNING_CODE.HSWW_CANCEL_OUTDOOR]: 16,
  [WARNING_CODE.HSWW_AMBER_INDOOR]: 8,
  [WARNING_CODE.HSWW_RED_INDOOR]: 4,
  [WARNING_CODE.HSWW_BLACK_INDOOR]: 2,
  [WARNING_CODE.HSWW_CANCEL_INDOOR]: 1,
};

export enum WARNING_STATUS {
  ISSUE = 'ISSUE',
  CANCEL = 'CANCEL',
}